Reorganize client file structure, update README
