+++ client/src/main.tsx
@@ -2,11 +2,11 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 
-import Root from './routes/Root.tsx';
-import Home from './routes/Home.tsx';
-import About from './routes/About';
-import Help from './routes/Help';
-import ErrorPage from './pages/ErrorPage.tsx';
+import Root from './pages/Root';
+import Home from './pages/Home';
+import About from './pages/About';
+import Help from './pages/Help';
+import ErrorPage from './pages/ErrorPage';
 import './index.css';
 
 const router = createBrowserRouter([
