+++ client/src/pages/Help.tsx
